chore: new other build files
creates two new builds for 512 bytes and 256 bytes. Also runs wit calc in the CI tests whereas before it was just being installed in the artifacts jobs and being used. I double checked the artifact Job and the make file and believe that they will both correctly make separate zips for each build. There was also one miscellaneous empty file with a .built extension that i assumed was a mistake so i removed it.